# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Haus of Wellness API
  version: 1.0.0
  description: Multi-tenant wellness platform API
servers:
  - url: /api/v1
paths:
...
  /organizations/{orgId}/bookings:
    get:
      summary: List bookings
      operationId: listBookings
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
        - name: filter[status]
          in: query
          schema:
            type: string
        - name: page[cursor]
          in: query
          schema:
            type: string
        - name: page[limit]
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
    post:
      summary: Create booking
      operationId: createBooking
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /organizations/{orgId}/bookings/{bookingId}:
    patch:
      summary: Update booking
      operationId: updateBooking
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /organizations/{orgId}/availability/check:
    post:
      summary: Check staff availability
      operationId: checkAvailability
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityCheckRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityCheckResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: haus_session
  parameters:
    orgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        staffId:
          type: string
          format: uuid
          nullable: true
        bookingDate:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
        isWalkin:
          type: boolean
    BookingListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        meta:
          type: object
          properties:
            nextCursor:
              type: string
              nullable: true
    CreateBookingRequest:
      type: object
      required: [customerId, bookingDate, startTime, endTime]
      properties:
        customerId:
          type: string
          format: uuid
        staffId:
          type: string
          format: uuid
        bookingDate:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        serviceIds:
          type: array
          items:
            type: string
            format: uuid
        isWalkin:
          type: boolean
    UpdateBookingRequest:
      type: object
      properties:
        status:
          type: string
        staffId:
          type: string
          format: uuid
    AvailabilityCheckRequest:
      type: object
      required: [staffId, bookingDate, startTime, endTime]
      properties:
        staffId:
          type: string
          format: uuid
        bookingDate:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        excludeBookingId:
          type: string
          format: uuid
    AvailabilityCheckResponse:
      type: object
      properties:
        available:
          type: boolean